Latest Code Commit
Task :23-09-2021

Working Time : 08.00am to 05.00pm

* Pet lover(Overall): Image crop for taking pics through the camera and selecting images from gallery -- *Completed*

* Doctor(Overall): Image crop for taking pics through the camera and selecting images from gallery and PDF file size 2mb greater than restrict (Bug Fix) -- *Completed*

* Service Provider(Overall): Image crop for taking pics through the camera and selecting images from gallery and PDF file size 2mb greater than restrict -- *Completed*

* Vendor(Overall): Image crop for taking pics through the camera and selecting images from gallery and PDF file size 2mb greater than restrict -- *Completed*

* Null data handling in the Doctor, Service Provider and Vendor -- *Completed*

+++ app/src/main/java/com/petfolio/infinituss/responsepojo/DoctorSearchResponse.java
@@ -94,7 +94,7 @@ public class DoctorSearchResponse {
         private String clinic_loc;
         private String communication_type;
         private String distance;
-        private double star_count;
+        private int star_count;
         private int review_count;
         private int doctor_exp;
         private String city_name;
@@ -185,11 +185,11 @@ public class DoctorSearchResponse {
             this.distance = distance;
         }
 
-        public double getStar_count() {
+        public int getStar_count() {
             return star_count;
         }
 
-        public void setStar_count(double star_count) {
+        public void setStar_count(int star_count) {
             this.star_count = star_count;
         }
